requirements.txt
cycler==0.10.0
decorator==4.4.1
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
Pillow==6.2.1
pycairo==1.13.1
pyclustering==0.9.2
pyparsing==2.4.4
python-dateutil==2.8.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
sklearn==0.0
snap-stanford==5.0.0
tqdm==4.37.0
